relocate test/__lib__ -> test/lib
