+++ src/cmd/whoami.test.js
@@ -1,6 +1,6 @@
 const proxyquire = require('proxyquire');
 const { expect, sinon } = require('../../test/setup');
-const { withConsoleStubs } = require('../../test/__lib__/mocha-utils');
+const { withConsoleStubs } = require('../../test/lib/mocha-utils');
 
 const stubs = {
 	api: {
