+++ test/e2e/binary.e2e.js
@@ -1,9 +1,9 @@
 const path = require('path');
 const { expect } = require('../setup');
-const cli = require('../__lib__/cli');
+const cli = require('../lib/cli');
 const {
 	PATH_FIXTURES_BINARIES_DIR
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Binary Command', () => {
