+++ test/e2e/call.e2e.js
@@ -1,8 +1,8 @@
 const { expect } = require('../setup');
-const cli = require('../__lib__/cli');
+const cli = require('../lib/cli');
 const {
 	DEVICE_NAME
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Call Command [@device]', () => {
