+++ test/e2e/cloud.e2e.js
@@ -1,14 +1,14 @@
 const path = require('path');
 const capitalize = require('lodash/capitalize');
 const { expect } = require('../setup');
-const cli = require('../__lib__/cli');
+const cli = require('../lib/cli');
 const {
 	DEVICE_ID,
 	DEVICE_NAME,
 	DEVICE_PLATFORM_NAME,
 	PATH_TMP_DIR,
 	PATH_PROJ_STROBY_INO
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Cloud Command [@device]', () => {
