+++ test/e2e/config.e2e.js
@@ -1,10 +1,10 @@
 const path = require('path');
 const { expect } = require('../setup');
-const cli = require('../__lib__/cli');
-const fs = require('../__lib__/fs');
+const cli = require('../lib/cli');
+const fs = require('../lib/fs');
 const {
 	PATH_PARTICLE_DIR
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Config Command', () => {
