+++ test/e2e/list.e2e.js
@@ -1,11 +1,11 @@
 const capitalize = require('lodash/capitalize');
 const { expect } = require('../setup');
-const cli = require('../__lib__/cli');
+const cli = require('../lib/cli');
 const {
 	DEVICE_ID,
 	DEVICE_NAME,
 	DEVICE_PLATFORM_NAME
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('List Commands', () => {
