+++ test/e2e/login.e2e.js
@@ -1,10 +1,10 @@
 const { expect } = require('../setup');
-const { delay } = require('../__lib__/mocha-utils');
-const cli = require('../__lib__/cli');
+const { delay } = require('../lib/mocha-utils');
+const cli = require('../lib/cli');
 const {
 	USERNAME,
 	PASSWORD
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Login Command', () => {
