+++ test/e2e/logout.e2e.js
@@ -1,9 +1,9 @@
 const { expect } = require('../setup');
-const { delay } = require('../__lib__/mocha-utils');
-const cli = require('../__lib__/cli');
+const { delay } = require('../lib/mocha-utils');
+const cli = require('../lib/cli');
 const {
 	USERNAME
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Logout Command', () => {
