+++ test/e2e/serial.e2e.js
@@ -1,13 +1,13 @@
 const words = require('lodash/words');
 const capitalize = require('lodash/capitalize');
 const { expect } = require('../setup');
-const { runForAtLeast } = require('../__lib__/mocha-utils');
-const cli = require('../__lib__/cli');
+const { runForAtLeast } = require('../lib/mocha-utils');
+const cli = require('../lib/cli');
 const {
 	DEVICE_ID,
 	DEVICE_PLATFORM_ID,
 	DEVICE_PLATFORM_NAME
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Serial Commands [@device]', () => {
