+++ test/e2e/token.e2e.js
@@ -1,11 +1,11 @@
 const { expect } = require('../setup');
-const { delay } = require('../__lib__/mocha-utils');
-const matches = require('../__lib__/capture-matches');
-const stripANSI = require('../__lib__/ansi-strip');
-const cli = require('../__lib__/cli');
+const { delay } = require('../lib/mocha-utils');
+const matches = require('../lib/capture-matches');
+const stripANSI = require('../lib/ansi-strip');
+const cli = require('../lib/cli');
 const {
 	PASSWORD
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Token Commands', () => {
