+++ test/e2e/udp.e2e.js
@@ -1,6 +1,6 @@
 const { expect } = require('../setup');
-const { delay } = require('../__lib__/mocha-utils');
-const cli = require('../__lib__/cli');
+const { delay } = require('../lib/mocha-utils');
+const cli = require('../lib/cli');
 
 
 describe('UDP Commands', () => {
