+++ test/e2e/variable.e2e.js
@@ -1,10 +1,10 @@
 const { expect } = require('../setup');
-const { delay } = require('../__lib__/mocha-utils');
-const cli = require('../__lib__/cli');
+const { delay } = require('../lib/mocha-utils');
+const cli = require('../lib/cli');
 const {
 	DEVICE_ID,
 	DEVICE_NAME
-} = require('../__lib__/env');
+} = require('../lib/env');
 
 
 describe('Variable Commands [@device]', () => {
