+++ test/lib/ansi-strip.js
