+++ test/lib/capture-matches.js
