+++ test/lib/cli.js
