+++ test/lib/fs.js
